% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squash.r
\name{squash}
\alias{squash}
\alias{squash0}
\title{Squash a list to a depth of 1}
\usage{
squash(x, keep.empty = FALSE)

squash0(x)
}
\arguments{
\item{x}{A list.}

\item{keep.empty}{Logical, indicating whether to retain
empty elements (e.g., \code{NULL}, \code{numeric(0)}, etc.) in the
squashed list. Defaults to \code{FALSE}.}
}
\value{
A list of depth 1. If a non-list (or data.frame)
is provided, the input is returned. If the squashed list
is empty, \code{NULL} is returned.
}
\description{
Squash a nested list into a list of depth 1.
}
\details{
\code{squash()} differs to \code{\link[=unlist]{unlist()}} in three key ways:
\enumerate{
\item It does not simplify the result, \strong{nor} any elements
of the input list.
\item data.frame elements are left as is, rather than
being unlisted/squashed.
\item Names are not combined, inner names are preserved.
}

\code{squash0()} is a barebones version of \code{squash()}, which
strips names and drops empty elements by default for a
slight performance gain.
}
\examples{
# no simplification, always 1d list result
squash(list(x = 1, list(a = 2, 3)))
squash0(list(x = 1, list(a = 2, 3)))

# data.frames left as is
squash(list(x = 1, y = data.frame(a = 2, b = 3)))
squash0(list(x = 1, y = data.frame(a = 2, b = 3)))

# empty list returns NULL
squash(list())
squash0(list(NULL, numeric()))

# inner names are preserved
squash(list(a = 1, b = list(list(a = 1))))

# result nor any elements of the input list are simplified
unlist(list(1, 2, 3))
squash0(list(1, 2, 3))
unlist(list(1:3, list("a", "b")), recursive = FALSE)
squash0(list(1:3, list("a", "b")))
}
\seealso{
\code{\link[=squash_track]{squash_track()}} for squashing whilst encoding the
original structure through names.
}
