% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squash_track.r
\name{squash_track}
\alias{squash_track}
\alias{squash_track.list}
\alias{squash_track0}
\alias{squash_track0.list}
\alias{is.squashed}
\title{Squash a list to a depth of 1 whilst encoding structure in names}
\usage{
squash_track(x, ...)

\method{squash_track}{list}(x, sep = "..", wrap = "'", unique_names = FALSE, ...)

squash_track0(x, ...)

\method{squash_track0}{list}(x, ...)

is.squashed(x)
}
\arguments{
\item{x}{A list or data.frame. For \code{is.squashed()}, an object to check.}

\item{...}{additional arguments to pass to methods.}

\item{sep}{String separator to use between levels in the names.}

\item{wrap}{String to wrap around whole number names.}

\item{unique_names}{Either \code{FALSE} to allow duplicate names, or a string
separator to be placed in between the name and the index to differentiate.}
}
\value{
For \code{squash_track()}, a squashed, 1-level list with class \code{squashed} and the
\code{sep} and \code{wrap} inputs attached as attributes. \code{unique_names} is also an
attribute if not \code{FALSE}.

For \code{squash_track0()}, a squashed, 1-level list with class \code{squashed0}.

For \code{is.squashed()}, \code{TRUE} or \code{FALSE}.
}
\description{
\code{squash_track()}  and \code{squash_track0()} reduce a list to a single level,
tracking and encoding the original structure in the names of the
elements. \code{squash_track0()} errors on duplicate names at a given
level, whereas \code{squash_track()} differentiates names flexibly using
given \code{sep}, \code{wrap}, and \code{unique_names} parameters.
}
\details{
To encode the original structure, \code{squash_track()} concatenates the
names of the elements using the specified \code{sep}. If an element is unnamed
the index is used instead. If an element is named and the name is numeric
and whole (aka could be confused for an index), then it is wrapped with
the specified \code{wrap} string. The resulting name structure can therefore
be a mix of names and indices.

If \code{unique_names} is a string, duplicate names will be made unique by
appending the index to the end of the name, separated by the string given.

To ensure the original structure can be reconstructed, any names that match
the \code{squash} patterns upon ingest will cause an error: \code{sep} cannot be
present in any of the element names; no element names can already be a
whole number with the \code{wrap} around it; and if \code{unique_names} is
a string, then no element names can already end with the \code{unique_names}
separator followed by a whole number.

\link{data.frame} elements of input lists, as well as empty list elements,
are left as is. This is \strong{unlike} the default behaviour of \code{squash()}
which drops empty elements.

\code{is.squashed()} checks if a list has been squashed by either \code{squash_track()}
or \code{squash_track0()}.
}
\examples{
# for named elements, names are concatenated with `sep`.
x <- list(a = list(b = 1, c = 2), d = 3)
squash_track(x)
squash_track0(x)

# for unnamed elements, indices are used
squash_track(list(list(1, 2), 3))
squash_track0(list(list(1, 2), 3))

# for mixed named/unnamed elements, names and indices are combined.
x <- list(a = list(1, 2), list(3, b = 4))
squash_track(x)
squash_track0(x)

# names that could also be indices are wrapped with `wrap`.
x <- list(10, "1" = 15, "2" = list("2" = 1, 5, "1.5" = 7))
squash_track(x)
squash_track0(x)

# duplicate names remain when `unique_names` is `FALSE`.
squash_track(list(a = 1, a = 2))
# these can be made unique using the index and a given string:
squash_track(list(a = 1, a = 2), unique_names = "*")
# fails for squash_track0
try(squash_track0(list(a = 1, a = 2)))

is.squashed(squash_track(list(a = 1)))
is.squashed(squash_track0(list(a = 1)))
}
\seealso{
\link{squash} to reduce a list to 1-d without tracking the original structure.

\link{squashed_nm2call} to convert a squashed name back to the original nested
structure as a call.
}
