% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm2call.r
\name{squashed_nm2call}
\alias{squashed_nm2call}
\alias{squashed0_nm2call}
\title{Convert squashed names to calls}
\usage{
squashed_nm2call(
  nm,
  var = ".data",
  sep = "..",
  wrap = "'",
  unique_names = FALSE
)

squashed0_nm2call(nm, var = ".data")
}
\arguments{
\item{nm}{A squashed name to be converted back to a call.}

\item{var}{A string referring to the object that will be the subject
of the call.}

\item{sep}{The separator used in the squashed names.}

\item{wrap}{The character used to wrap names in the squashed names.}

\item{unique_names}{The marker used for uniquely marked names.}
}
\value{
A call representing the original nested structure
of the squashed name.
}
\description{
Convert a squashed name back to its original
nested structure as a call.
}
\details{
For \code{squashed0_nm2call()} inputs should be derived from the
names of squashed0 objects and therefore will already be in
a format suitable for direct conversion back to calls using
\link{parse}.

\code{squashed_nm2call()} converts squashed names back to calls,
using the \code{sep}, \code{wrap}, and \code{unique_names} arguments to
correctly interpret the structure of the squashed names in
relation to the original object. For example, \code{"1..'1'"}
with \code{sep=".."} and \code{wrap="'"} would be interpreted as
\code{.data[[1]][["1"]]} whereas \verb{"}a..b..'1'..a*5\verb{"} would
be interpreted as \code{.data[["a"]][["b"]][["1"]][[5L]]}.
}
\examples{
squashed_nm2call("1..'1'")
squashed_nm2call("a..b..'1'..a*5", unique_names = "*")

x <- list(list("1" = 1))
y <- squash0(x)
cll <- squashed0_nm2call(names(y), var = "x")
eval(cll)
}
